Enum ggbc::ir::opcodes::Statement

pub enum Statement {
    Nop(usize),
    Stop(StopStatus),
    Ld {
        source: Source<u8>,
        destination: Destination,
    },
    LdW {
        source: Source<u16>,
        destination: Destination,
    },
    LdAddr {
        source: Source<Address>,
        destination: Destination,
    },
    Inc {
        source: Source<u8>,
        destination: Destination,
    },
    Dec {
        source: Source<u8>,
        destination: Destination,
    },
    IncW {
        source: Source<u16>,
        destination: Destination,
    },
    DecW {
        source: Source<u16>,
        destination: Destination,
    },
    Add {
        left: Source<u8>,
        right: Source<u8>,
        destination: Destination,
    },
    Sub {
        left: Source<u8>,
        right: Source<u8>,
        destination: Destination,
    },
    And {
        left: Source<u8>,
        right: Source<u8>,
        destination: Destination,
    },
    Xor {
        left: Source<u8>,
        right: Source<u8>,
        destination: Destination,
    },
    Or {
        left: Source<u8>,
        right: Source<u8>,
        destination: Destination,
    },
    LeftShift {
        left: Source<u8>,
        right: Source<u8>,
        destination: Destination,
    },
    RightShift {
        left: Source<u8>,
        right: Source<u8>,
        destination: Destination,
    },
    Mul {
        left: Source<u8>,
        right: Source<u8>,
        destination: Destination,
    },
    Div {
        left: Source<u8>,
        right: Source<u8>,
        destination: Destination,
    },
    Rem {
        left: Source<u8>,
        right: Source<u8>,
        destination: Destination,
    },
    AddW {
        left: Source<u16>,
        right: Source<u16>,
        destination: Destination,
    },
    SubW {
        left: Source<u16>,
        right: Source<u16>,
        destination: Destination,
    },
    AndW {
        left: Source<u16>,
        right: Source<u16>,
        destination: Destination,
    },
    XorW {
        left: Source<u16>,
        right: Source<u16>,
        destination: Destination,
    },
    OrW {
        left: Source<u16>,
        right: Source<u16>,
        destination: Destination,
    },
    LeftShiftW {
        left: Source<u16>,
        right: Source<u8>,
        destination: Destination,
    },
    RightShiftW {
        left: Source<u16>,
        right: Source<u8>,
        destination: Destination,
    },
    MulW {
        left: Source<u16>,
        right: Source<u16>,
        destination: Destination,
    },
    DivW {
        left: Source<u16>,
        right: Source<u16>,
        destination: Destination,
    },
    RemW {
        left: Source<u16>,
        right: Source<u16>,
        destination: Destination,
    },
    Eq {
        left: Source<u8>,
        right: Source<u8>,
        destination: Destination,
    },
    NotEq {
        left: Source<u8>,
        right: Source<u8>,
        destination: Destination,
    },
    Greater {
        left: Source<u8>,
        right: Source<u8>,
        destination: Destination,
    },
    GreaterEq {
        left: Source<u8>,
        right: Source<u8>,
        destination: Destination,
    },
    Less {
        left: Source<u8>,
        right: Source<u8>,
        destination: Destination,
    },
    LessEq {
        left: Source<u8>,
        right: Source<u8>,
        destination: Destination,
    },
    Jmp {
        location: Location,
    },
    JmpCmp {
        location: Location,
        source: Source<u8>,
    },
    JmpCmpNot {
        location: Location,
        source: Source<u8>,
    },
    Call {
        routine: usize,
        range: Range<u16>,
    },
    Ret,
}

The statements, or instruction set of the IR.

Variants

Nop(usize)

Do nothing, used as placeholder.

Stop(StopStatus)

Statement to stop execution (end program).

Ld

8bit load.

Fields of Ld

source: Source<u8>destination: Destination

LdW

16bit load.

Fields of LdW

source: Source<u16>destination: Destination

LdAddr

Load address.

Fields of LdAddr

source: Source<Address>destination: Destination

Inc

8bit increment.

Fields of Inc

source: Source<u8>destination: Destination

Dec

8bit decrement.

Fields of Dec

source: Source<u8>destination: Destination

IncW

16bit increment.

Fields of IncW

source: Source<u16>destination: Destination

DecW

16bit decrement.

Fields of DecW

source: Source<u16>destination: Destination

Add

8bit add.

Fields of Add

left: Source<u8>right: Source<u8>destination: Destination

Sub

8bit subtract.

Fields of Sub

left: Source<u8>right: Source<u8>destination: Destination

And

8bit bitwise AND.

Fields of And

left: Source<u8>right: Source<u8>destination: Destination

Xor

8bit bitwise XOR.

Fields of Xor

left: Source<u8>right: Source<u8>destination: Destination

Or

8bit bitwise OR.

Fields of Or

left: Source<u8>right: Source<u8>destination: Destination

LeftShift

8bit left shift.

Fields of LeftShift

left: Source<u8>right: Source<u8>destination: Destination

RightShift

8bit right shift.

Fields of RightShift

left: Source<u8>right: Source<u8>destination: Destination

Mul

8bit multiply.

Fields of Mul

left: Source<u8>right: Source<u8>destination: Destination

Div

8bit divide.

Fields of Div

left: Source<u8>right: Source<u8>destination: Destination

Rem

8bit remainder.

Fields of Rem

left: Source<u8>right: Source<u8>destination: Destination

AddW

16bit add.

Fields of AddW

left: Source<u16>right: Source<u16>destination: Destination

SubW

16bit subtract.

Fields of SubW

left: Source<u16>right: Source<u16>destination: Destination

AndW

16bit bitwise AND.

Fields of AndW

left: Source<u16>right: Source<u16>destination: Destination

XorW

16bit bitwise XOR.

Fields of XorW

left: Source<u16>right: Source<u16>destination: Destination

OrW

16bit bitwise OR.

Fields of OrW

left: Source<u16>right: Source<u16>destination: Destination

LeftShiftW

16bit left shift.

Fields of LeftShiftW

left: Source<u16>right: Source<u8>destination: Destination

RightShiftW

16bit right shift.

Fields of RightShiftW

left: Source<u16>right: Source<u8>destination: Destination

MulW

16bit multiply.

Fields of MulW

left: Source<u16>right: Source<u16>destination: Destination

DivW

16bit divide.

Fields of DivW

left: Source<u16>right: Source<u16>destination: Destination

RemW

16bit remainder.

Fields of RemW

left: Source<u16>right: Source<u16>destination: Destination

Eq

8bit boolean equals.

Fields of Eq

left: Source<u8>right: Source<u8>destination: Destination

NotEq

8bit boolean not-equals.

Fields of NotEq

left: Source<u8>right: Source<u8>destination: Destination

Greater

8bit boolean greater-than.

Fields of Greater

left: Source<u8>right: Source<u8>destination: Destination

GreaterEq

8bit boolean greater-or-equal-than.

Fields of GreaterEq

left: Source<u8>right: Source<u8>destination: Destination

Less

8bit boolean less-than.

Fields of Less

left: Source<u8>right: Source<u8>destination: Destination

LessEq

8bit boolean less-or-equal-than.

Fields of LessEq

left: Source<u8>right: Source<u8>destination: Destination

Jmp

Jump to location.

Fields of Jmp

location: Location

JmpCmp

Conditional jump to location. Jumps to the given location if source resolves to zero.

Fields of JmpCmp

location: Locationsource: Source<u8>

JmpCmpNot

Conditional jump Jumps to the given location if source resolves to non-zero.

Fields of JmpCmpNot

location: Locationsource: Source<u8>

Call

Routine call.

Fields of Call

routine: usize

Routine index.

range: Range<u16>

range of the current stack frame corresponding to the beginning of the new function's stack frame.

Ret

Return from routine.

Implementations

impl Statement

pub fn display(&self) -> Mnemonic<'_>

Return displayable mnemonic.